Fix another warning detected by Visual Studio (negating an unsigned int).

+++ gs/base/gxfcache.h
@@ -234,7 +234,7 @@ struct cached_char_s {
  */
 #define align_cached_char_mod align_cached_bits_mod
 #define sizeof_cached_char\
-  ROUND_UP(sizeof(cached_char), align_cached_char_mod)
+  ROUND_UP((int)sizeof(cached_char), align_cached_char_mod)
 #define cc_bits(cc) ((byte *)(cc) + sizeof_cached_char)
 #define cc_const_bits(cc) ((const byte *)(cc) + sizeof_cached_char)
